nameに関するバリデーション実装
