+++ memo/validation-memo.md
@@ -55,3 +55,7 @@
 
 - STRICT:日付と時刻を厳密に解決する
 - LENIENT:日付けと時刻を厳密ではない方法で解決する
+
+### エラー対策
+
+- 期待値と違うエラー: デバッグして`response`を確認する
